# specs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Servicing API",
    "description": "API to get service details about customers"
  },
  "servers": [
    {
      "url": "host"
    }
  ],
...
  "paths": {
...
    "/customer_accounts/{id}/service_accounts": {
      "get": {
        "tags": [
          "customer_accounts",
          "service_accounts"
        ],
        "summary": "Retrieves service accounts for a customer id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of customer account to get service accounts for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/service_account"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No Service Accounts not found"
          }
        }
      }
    },
...
    "/service_accounts/{id}": {
      "get": {
        "tags": [
          "service_accounts"
        ],
        "summary": "Retrieves service account information by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/service_account"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      }
    },
    "/service_accounts/{id}/autopay": {
      "get": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Retrieves autopay details service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch autopay details",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Id": {
                      "type": "string"
                    },
                    "StartDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "StopDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "PayMethodId": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      },
      "post": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Sign up for autopay",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to sign up in autopay",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/autopay_request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Id": {
                      "type": "string"
                    },
                    "StartDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "StopDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "PayMethodId": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      },
      "delete": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Cancels autopay for service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to cancel autopay for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autopay canceled"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account or autopay not found"
          }
        }
      }
    },
    "/service_accounts/{id}/budget_billing": {
      "get": {
        "tags": [
          "service_accounts",
          "budget_billing"
        ],
        "summary": "Retrieves budget billing information by service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch budget billing info for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns budget billing information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Indicator": {
                        "type": "boolean"
                      },
                      "Amount": {
                        "type": "number"
                      },
                      "Create_Date": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "Variance": {
                        "type": "number"
                      },
                      "Pending_Cancel_Count": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      },
      "post": {
        "tags": [
          "service_accounts",
          "budget_billing"
        ],
        "summary": "Signs up for budget billing for service account specified",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to sign up on budget billing for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Amount": {
                    "type": "number"
                  }
                },
                "required": [
                  "Amount"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a budget billing confirmation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      },
      "delete": {
        "tags": [
          "service_accounts",
          "budget_billing"
        ],
        "summary": "Cancels budget billing for a service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to cancel budget billing for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      }
    },
...
    "/service_accounts/{id}/invoices": {
      "get": {
        "tags": [
          "service_accounts",
          "invoices"
        ],
        "summary": "Retrieves invoices by service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch invoices for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a list of invoices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/invoice"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account or Invoices Not Found"
          }
        }
      }
    },
...
    "/service_accounts/{id}/Meter_Read_Details/{count}": {
      "get": {
        "tags": [
          "service_accounts"
        ],
        "summary": "Retrieves meter read details for a service account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to search by",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "in": "path",
            "description": "count of meter read details to be retrieved",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account meter read details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/invoice_meter_read_detail"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      }
    },
...
    "/service_accounts/{id}/notes": {
      "post": {
        "tags": [
          "service_accounts",
          "notes"
        ],
        "summary": "save service account notes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to add note to",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/csp_notes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a service account notes response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      }
    },
    "/service_accounts/{id}/payment_extension": {
      "get": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Retrieves payment extension status by service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch payment extension status for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/payment_extension_status"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Requests for a payment extension towards the service account specified.  Will be set to 7 days past the action date unless extend is set to true in the request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to request a payment extension for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/payment_extension_request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/payment_extension_response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/payment_extension_status_code"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/service_accounts/{id}/payments": {
      "get": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Retrieves payments by service account id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch payments for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/payment_detail"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      },
      "post": {
        "tags": [
          "service_accounts",
          "payments"
        ],
        "summary": "Makes a payment towards the service account specified",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to make a payment for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/payment_request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a service account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/payment_confirmation"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      }
    },
    "/service_accounts/{id}/service_orders": {
      "get": {
        "tags": [
          "service_accounts"
        ],
        "summary": "Gets the list of service orders for a service account.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns service orders for a service account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Service_Order_Description": {
                        "type": "string"
                      },
                      "Date_Processed": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "Distributor_Name": {
                        "type": "string"
                      },
                      "Priority_Code": {
                        "type": "string"
                      }
                    }
                  }
                },
                "example": [
                  {
                    "Service_Order_Description": "Reconnect Complete",
                    "Date_Processed": "05/02/2019",
                    "Distributor_Name": "AEP - Texas Central",
                    "priority_code": ""
                  },
                  {
                    "Service_Order_Description": "Initiate Reconnect",
                    "Date_Processed": "05/02/2019",
                    "Distributor_Name": "AEP - Texas Central",
                    "priority_code": "01"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      }
    },
    "/service_accounts/{id}/standard_waivers": {
      "get": {
        "tags": [
          "service_accounts"
        ],
        "summary": "Retrieves standard waivers for a service account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of service account to get waivers",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 returns service account waivers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "FeeCreationDate": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "FeeType": {
                        "type": "string",
                        "enum": [
                          "Billed",
                          "ManualPending",
                          "ServicePending"
                        ]
                      },
                      "Fee": {
                        "type": "string"
                      },
                      "Tax": {
                        "type": "string"
                      },
                      "WaiveStatus": {
                        "type": "string",
                        "enum": [
                          "Billed",
                          "Approved",
                          "Rejected",
                          "Pending",
                          "None"
                        ]
                      },
                      "Description": {
                        "type": "string"
                      },
                      "FeeWaiverType": {
                        "type": "string",
                        "enum": [
                          "Late_Fee",
                          "NSF_Fees",
                          "Disconnect_Fees",
                          "Reconnect_Fees",
                          "Disconnect_Letter_Fees"
                        ]
                      }
                    }
                  }
                },
                "example": [
                  {
                    "FeeCreationDate": "06/14/2019",
                    "FeeType": "ServicePending",
                    "Fee": 6.82,
                    "Tax": 0.14756434,
                    "WaiveStatus": "None",
                    "Description": "",
                    "FeeWaiverType": "Late_Fee"
                  },
                  {
                    "FeeCreationDate": "03/31/2019",
                    "FeeType": "ServicePending",
                    "Fee": 7.29,
                    "Tax": 0.15773373,
                    "WaiveStatus": "None",
                    "Description": "",
                    "FeeWaiverType": "Late_Fee"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      },
      "post": {
        "tags": [
          "service_accounts"
        ],
        "summary": "Requests a waiver for the service account.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the service account for which the waiver is requested.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/waiver_request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 returns a success flag",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Is_Success": {
                      "type": "boolean"
                    },
                    "Description": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "Is_Success": true,
                  "Description": "test"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service Account not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "autopay_request": {
        "type": "object",
        "properties": {
          "pay_method_id": {
            "type": "string"
          }
        }
      },
...
      "error_list": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
      "invoice": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string"
          },
          "Invoice_Date": {
            "type": "string",
            "format": "date"
          },
          "Due_Date": {
            "type": "string",
            "format": "date"
          },
          "Usage": {
            "type": "integer"
          },
          "Current_Charges": {
            "type": "number"
          },
          "Amount_Due": {
            "type": "number"
          },
          "Credit": {
            "type": "number"
          },
          "Balance_Forward": {
            "type": "number"
          }
        }
      },
...
      "invoice_meter_read_detail": {
        "type": "object",
        "properties": {
          "Invoice_Id": {
            "type": "string"
          },
          "From_Date": {
            "type": "string",
            "format": "date"
          },
          "To_Date": {
            "type": "string",
            "format": "date"
          },
          "Read_Type": {
            "type": "string"
          },
          "From_Read": {
            "type": "string"
          },
          "To_Read": {
            "type": "string"
          },
          "Usage": {
            "type": "string"
          }
        }
      },
...
      "payment_extension_response": {
        "type": "object",
        "properties": {
          "Extended_Date": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "payment_extension_request": {
        "type": "object",
        "properties": {
          "Extend": {
            "description": "Flag that indicates that the payment extension will be extended an extra 3 days past the normal 7",
            "type": "boolean"
          }
        }
      },
...
      "payment_extension_status_code": {
        "type": "string",
        "enum": [
          "UNKNOWN",
          "UNSUCCESSFUL",
          "NO_PAST_DUE_FOUND",
          "ALREADY_APPROVED",
          "DEPOSIT_CHARGE_FOUND",
          "FEWER_BILL_COUNT_FOUND",
          "NO_DISCONNECT_LETTER",
          "BALANCE_TRANSFER_FOUND",
          "NFS_FOUND",
          "DISCONNECT_ORDER_FOUND",
          "BROKEN_ARRANGEMENT_FOUND",
          "UNPAID_PAST_DUE",
          "UNPAID_ARRANGEMENT_DUE",
          "NONE_PAID_UNTIL_FOLLOW_UP_DATE",
          "EXCESSIVE_BROKEN_ARRANGEMENT_FOUND",
          "NON_PAYMENT_ORDER_FOUND",
          "EXPIRED_DISCONNECT_DATE",
          "NO_ACTION_DATE_FOUND",
          "DISCONNECT_NOTICE_NOT_ISSUED",
          "INSERT_EXTENSION_NOTES_FAILED",
          "CSP_NOT_ACTIVE",
          "CSP_STOP_REQUEST_FOUND",
          "NO_SUSPENSION_LETTER_ISSUED",
          "SUSPENSION_ORDER_ISSUED",
          "NON_PAYMENT_ORDER_FOUND_IN_LAST_SIX_MONTHS"
        ]
      },
...
      "csp_notes": {
        "type": "object",
        "properties": {
          "Note": {
            "type": "string"
          },
          "Followup_Ind": {
            "type": "number"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}